Refactor project structure and version handling; remove core module, update imports, and clean up documentation.
