+++ docs/index.md
@@ -1,3 +1,1 @@
-# pytest-http-vcr
-
-A pytest plugin that records and replays all HTTP interactions using a VCR-style cassette and a local HTTP server.
+--8<-- "README.md"
