+++ pyproject.toml
@@ -34,7 +34,7 @@ requires = ["setuptools>=70.1"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.dynamic]
-version = {attr = "pytest_http_vcr.__version__"}
+version = {attr = "pytest_http_vcr._version.__version__"}
 
 [dependency-groups]
 dev = [
