+++ src/pytest_http_vcr/__init__.py
@@ -1,7 +1,6 @@
-"""A pytest plugin that records and replays all HTTP interactions using a VCR-style cassette and a local HTTP server."""
+"""pytest_http_vcr."""
 
+from pytest_http_vcr._version import __version__
 from pytest_http_vcr.cli import main
-
-from pytest_http_vcr.core import __version__
 
 __all__ = ["__version__", "main"]
