+++ src/pytest_http_vcr/_version.py
@@ -1,0 +1,3 @@
+"""Version information for pytest_http_vcr."""
+
+__version__ = "0.1.0"
