+++ src/pytest_http_vcr/cli.py
@@ -2,13 +2,14 @@
 
 from argparse import ArgumentParser
 
-from pytest_http_vcr.core import __version__
+from pytest_http_vcr._version import __version__
 
 
 def generate_cli_parser() -> ArgumentParser:
     """Generate the argument parser for the pytest_http_vcr CLI."""
     parser = ArgumentParser(
-        description="A pytest plugin that records and replays all HTTP interactions using a VCR-style cassette and a local HTTP server."
+        description="A pytest plugin that records and replays all "
+        "HTTP interactions using a VCR-style cassette and a local HTTP server."
     )
     parser.add_argument("--version", action="version", version=__version__)
     return parser
