--- src/pytest_http_vcr/core.py
@@ -1,3 +1,0 @@
-"""Core package for pytest_http_vcr."""
-
-__version__ = "0.1.0"
